+++ packages/e2e-tests/test/LocalMerkle.test.ts
@@ -2,40 +2,65 @@ import { expect } from "chai";
 import { ethers } from "hardhat";
 import * as fs from "fs";
 import {
-  OffchainMerkleTree,
-  OffchainMerkleTree__factory,
+  Wallet,
+  Wallet__factory,
   TestSubtreeUpdateVerifier__factory,
+  Spend2Verifier__factory,
+  Vault__factory,
+  Vault,
+  SimpleERC20Token__factory
 } from "@flax/contracts";
 import {
   LocalMerkleProver,
   FlaxLMDB,
   DEFAULT_DB_PATH,
-  BinaryPoseidonTree
+  BinaryPoseidonTree,
+  FlaxPrivKey,
+  FlaxSigner,
 } from "@flax/sdk";
-import { fillBatch } from "./utils";
+import { SimpleERC20Token } from "@flax/contracts/dist/src/SimpleERC20Token";
+import { depositFunds } from "./utils";
 
 describe("LocalMerkle", async () => {
   let deployer: ethers.Signer;
-  let merkle: OffchainMerkleTree;
+  let alice: ethers.Signer;
+  let wallet: Wallet;
+  let token: SimpleERC20Token;
+  let vault: Vault;
   let db: FlaxLMDB;
   let localMerkle: LocalMerkleProver;
+  let flaxSigner: FlaxSigner;
 
   async function setup() {
+    const sk = BigInt(1);
+    const flaxPrivKey = new FlaxPrivKey(sk);
+    flaxSigner = new FlaxSigner(flaxPrivKey);
+
     db = new FlaxLMDB({ dbPath: DEFAULT_DB_PATH, localMerkle: true });
 
-    [deployer] = await ethers.getSigners();
+    [deployer, alice] = await ethers.getSigners();
     const subtreeupdateVerifierFactory = new TestSubtreeUpdateVerifier__factory(deployer);
     const subtreeUpdateVerifier = await subtreeupdateVerifierFactory.deploy();
 
-    const merkleFactory = new OffchainMerkleTree__factory(deployer);
-    merkle = await merkleFactory.deploy(subtreeUpdateVerifier.address);
+    const spend2VerifierFactory = new Spend2Verifier__factory(deployer);
+    const spend2Verifier = await spend2VerifierFactory.deploy();
 
-    localMerkle = new LocalMerkleProver(merkle.address, ethers.provider, db);
+    const tokenFactory = new SimpleERC20Token__factory(deployer);
+    token = await tokenFactory.deploy();
+
+    const vaultFactory = new Vault__factory(deployer);
+    vault = await vaultFactory.deploy();
+
+    const walletFactory = new Wallet__factory(deployer);
+    wallet = await walletFactory.deploy(vault.address, spend2Verifier.address, subtreeUpdateVerifier.address);
+
+    await vault.initialize(wallet.address);
+    localMerkle = new LocalMerkleProver(wallet.address, ethers.provider, db);
   }
 
   async function applySubtreeUpdate() {
-	const root = localMerkle.root();
-	await merkle.applySubtreeUpdate(root, [0n, 0n, 0n, 0n, 0n, 0n, 0n, 0n]);
+    const root = localMerkle.root();
+    await wallet.applySubtreeUpdate(root, [0n, 0n, 0n, 0n, 0n, 0n, 0n, 0n]);
   }
 
   beforeEach(async () => {
@@ -53,27 +78,34 @@ describe("LocalMerkle", async () => {
 
   it("Local merkle prover self syncs", async () => {
     console.log("Depositing 2 notes");
-    await merkle.insertNoteCommitments([0n, 1n]);
+    const ncs = await depositFunds(
+      wallet,
+      vault,
+      token,
+      alice,
+      flaxSigner.address,
+      [100n, 100n] 
+    );
 
     console.log("Fetching and storing leaves from events");
     await localMerkle.fetchLeavesAndUpdate();
     expect(localMerkle.count).to.eql(2);
 
     console.log("Ensure leaves match enqueued");
-    expect(BigInt(localMerkle.getProof(0).leaf)).to.equal(0n);
-    expect(BigInt(localMerkle.getProof(1).leaf)).to.equal(1n);
+    expect(BigInt(localMerkle.getProof(0).leaf)).to.equal(ncs[0]);
+    expect(BigInt(localMerkle.getProof(1).leaf)).to.equal(ncs[1]);
 
     console.log("filling subtree");
-    await fillBatch(merkle);
+    await wallet.fillBatchWithZeros();
 
     console.log("local merkle prover picks up the zeros")
     await localMerkle.fetchLeavesAndUpdate();
     expect(localMerkle.count).to.eql(BinaryPoseidonTree.BATCH_SIZE);
 
-
     console.log("Local merkle doesn't care when subtree update is applied")
     await applySubtreeUpdate();
-    expect(BigInt(localMerkle.getProof(0).leaf)).to.equal(0n);
-    expect(BigInt(localMerkle.getProof(1).leaf)).to.equal(1n);
+    expect(localMerkle.count).to.eql(BinaryPoseidonTree.BATCH_SIZE);
+    expect(BigInt(localMerkle.getProof(0).leaf)).to.equal(ncs[0]);
+    expect(BigInt(localMerkle.getProof(1).leaf)).to.equal(ncs[1]);
   });
 });