+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- move `OffchainMerkleTree` into a library and have `CommitmentTreeManager` hold the state
 - remove `IBatchMerkle` and `BatchBinaryMerkle`
 - split `Utils` into `TreeUtils` and `Utils`
 - Add fixture unit test for `SubtreeUpdateVerifier`